backend/requirements.txt
asgiref==3.7.2
Django==5.0.2
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.14.0
psycopg2-binary==2.9.9
pytz==2024.1
sqlparse==0.4.4
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
tensorflow==2.19.0
pandas>=2.2.0
scikit-learn>=1.4.0
numpy>=1.26.0
